python/requirements.txt
python-socketio
eventlet
flask
